the main idea of this PR is keeping cache locality for predicted rows.

Results:

Data set | predict time (master), s | predict time (this PR), s | Speedup
-- | -- | -- | --
 higgs1m | 4.12 | 1.32 | 3.12
 msrank |  13.9 | 6.07 |  2.29

without Dmatrix create:
Data set | predict time (master), s | predict time (this PR), s | Speedup
-- | -- | -- | --
 higgs1m | 4.12 - 0.25 | 1.32 - 0.25 | 3.62
 msrank |  13.9 - 2.6 | 6.07 - 2.6 |  3.25


Data set | logloss (master) | logloss (this PR) 
-- | -- | -- 
 higgs1m | 0.4036675 | 0.4036675 
 msrank |  0.8550638 | 0.8550638