@ShvetsKS 2020-10-01
@hcho3  not sure that failed steps are related to current changes
https://travis-ci.org/github/dmlc/xgboost/jobs/731946554#L8110