@hcho3 2020-10-04
I'll review today. Thanks for removing the unrolling variable kUnroll.