@ShvetsKS 2020-10-06
> I'll review today. Thanks for removing the unrolling variable kUnroll.
@hcho3  Thanks  for review in advance :)
With new approach there is no significant performance gain from unrolling.