@ShvetsKS 2020-10-07
@hcho3 Thanks a lot for your review.
this construction
` (*this)[pid].LeftChild() + !(fvalue < (*this)[pid].SplitCond());` 
brings up to 35% of speedup, it was reflected in comment: https://github.com/dmlc/xgboost/pull/6127/files#diff-d7a927cebbc6e99b6432fdba2b54231cR48

it helps to reduce miss branch prediction.