@hcho3 2020-10-07
Then please add a comment directly on top of the statement `(*this)[pid].LeftChild() + !(fvalue < (*this)[pid].SplitCond());` to explain what it does.