@ShvetsKS 2020-10-08
@hcho3 could take a look at failed tests, please
I think similar failed tests were discovered before the current PR.
Here [Doc] Add info on GPU compiler (#6204): https://xgboost-ci.net/blue/organizations/jenkins/xgboost/detail/master/525/pipeline fail is
similar to https://xgboost-ci.net/blue/organizations/jenkins/xgboost/detail/PR-6127/16/pipeline

should we restart it?